Changing double precision to single

+++ src/warpkit.cpp
@@ -6,15 +6,15 @@
 namespace py = pybind11;
 
 PYBIND11_MODULE(warpkit_cpp, m) {
-    py::class_<JuliaContext<double>>(m, "JuliaContext")
+    py::class_<JuliaContext<float>>(m, "JuliaContext")
         .def(py::init<>())
-        .def("romeo_voxelquality", &JuliaContext<double>::romeo_voxelquality, "Wrapper for ROMEO voxelquality function",
+        .def("romeo_voxelquality", &JuliaContext<float>::romeo_voxelquality, "Wrapper for ROMEO voxelquality function",
              py::arg("phase"), py::arg("TEs"), py::arg("mag"), py::return_value_policy::move)
-        .def("romeo_unwrap3D", &JuliaContext<double>::romeo_unwrap3D, "Wrapper for ROMEO unwrap_individual function",
+        .def("romeo_unwrap3D", &JuliaContext<float>::romeo_unwrap3D, "Wrapper for ROMEO unwrap_individual function",
              py::arg("phase"), py::arg("weights"), py::arg("mag"), py::arg("mask"), py::arg("correct_global") = true,
              py::arg("maxseeds") = 1, py::arg("merge_regions") = false, py::arg("correct_regions") = false,
              py::return_value_policy::move)
-        .def("romeo_unwrap4D", &JuliaContext<double>::romeo_unwrap4D, "Wrapper for ROMEO unwrap_individual function",
+        .def("romeo_unwrap4D", &JuliaContext<float>::romeo_unwrap4D, "Wrapper for ROMEO unwrap_individual function",
              py::arg("phase"), py::arg("TEs"), py::arg("weights"), py::arg("mag"), py::arg("mask"),
              py::arg("correct_global") = true, py::arg("maxseeds") = 1, py::arg("merge_regions") = false,
              py::arg("correct_regions") = false, py::return_value_policy::move);
